Fix narrowing conversions in generator and max() template deduction in ESP32 demos
- Updated generator/generate_tables.py to dynamically choose the correct type for GLYPH_BITMAPS (uint8_t, uint16_t, or uint32_t) based on font height.
- Fixed 'max' call in esp32_scenes.ino by casting to uint32_t to avoid conflicting types during template deduction.
- Updated all ESP32 demos to use uint32_t for colbyte when accessing GLYPH_BITMAPS to support larger font bitmaps.

+++ demo/esp32_scenes/arduino_tables.cpp
@@ -750,7 +750,7 @@ const uint8_t PROGMEM GLYPH_WIDTH = 20;
 const uint8_t PROGMEM GLYPH_HEIGHT = 19;
 const uint16_t PROGMEM GLYPH_COUNT = 39;
 const char PROGMEM GLYPH_CHAR_LIST[40] = " !?0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ";
-const uint8_t PROGMEM GLYPH_BITMAPS[780] = {
+const uint32_t PROGMEM GLYPH_BITMAPS[780] = {
   0, 0, 0, 0, 0, 0, 0, 0,
   0, 0, 0, 0, 0, 0, 0, 0,
   0, 0, 0, 0, 0, 0, 0, 425968,
